// src/app/modules/mail/mail.interface.ts
export type TMail = {
  fullName: string;
  email: string;
  dateRanges?: string;
  dateOfBirth?: string;
  timeOfBirth?: string;
  whatToLearn?: string;
  likeToFocus?: string;
  partnerName?: string;
  hopeToLearn?: string;
  placeOfBirth?: string;
  typeOfActivity?: string;
  partnerDateOfBirth?: string;
  partnerTimeOfBirth?: string;
  whatToLearnDetails?: string;
  partnerPlaceOfBirth?: string;
};